Add home page logic and connect to firebase database
Start new game button creates a new game in the database and add two player tokens (randomly generated) to this new game. The player is redirected to their player URL and the other player can connect using their player URL which matches their token.

Plan is to provide a link for player2 to join from.